refactor: make vehicle and circuit packages public
