+++ pkg/vehicles/vehicles_test.go
