+++ transformer_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/stretchr/testify/suite"
 	"github.com/zetetos/gt-telemetry/internal/telemetry"
-	"github.com/zetetos/gt-telemetry/internal/vehicles"
+	"github.com/zetetos/gt-telemetry/pkg/vehicles"
 )
 
 type TransformerTestSuite struct {
